+++ src/imio/helpers/xhtml.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from Acquisition import aq_base
+from html import escape
 from os import path
 from plone import api
 from plone.app.imaging.scale import ImageScale
@@ -429,13 +430,16 @@ def storeImagesLocally(context,
                        store_external_images=True,
                        store_internal_images=True,
                        pretty_print=False,
-                       force_resolve_uid=False):
+                       force_resolve_uid=False,
+                       replace_not_found_image=True):
     """If images are found in the given p_xhtmlContent,
        we download it and stored it in p_context, this way we ensure that it will
        always be available in case the external/internal image image disappear.
        If p_store_external_images is True, we retrieve external image and store it
        in p_context, if p_store_internal_images is True, we do the same for internal
-       images."""
+       images.
+       If p_replace_not_found_image is True, an image holding text "Image not available"
+       will be used if the image could not be retrieved."""
 
     def _handle_internal_image(img_src):
         """ """
@@ -524,10 +528,12 @@ def storeImagesLocally(context,
            ';base64,' not in original_img_src:
             continue
         filename = data = None
+        handled = False
         # external images
         if store_external_images and not original_img_src.startswith(portal_url) and \
            'resolveuid' not in original_img_src:
             filename, data = _handle_external_image(original_img_src)
+            handled = True
 
         # image in portal but not already stored in context
         # handle images using resolveuid
@@ -550,9 +556,20 @@ def storeImagesLocally(context,
            img_src.startswith(portal_url) and \
            not img_src.startswith(context_url):
             filename, data = _handle_internal_image(img_src)
+            handled = True
 
         if not filename:
-            continue
+            if handled:
+                if replace_not_found_image:
+                    filename = "imageNotFound.jpg"
+                    f = open(os.path.join(os.path.dirname(__file__), filename), 'r')
+                    data = f.read()
+                    f.close()
+                else:
+                    continue
+            else:
+                continue
+
         changed = True
 
         # create image
@@ -625,7 +642,7 @@ def separate_images(xhtmlContent, pretty_print=False):
                                        method='html') for x in tree.iterchildren()])
 
 
-def object_link(obj, view='view', attribute='Title', content='', target=''):
+def object_link(obj, view='view', attribute='Title', content='', target='', escaped=True):
     """ Returns an html link for the given object """
     href = view and "%s/%s" % (obj.absolute_url(), view) or obj.absolute_url()
     if not content:
@@ -636,4 +653,6 @@ def object_link(obj, view='view', attribute='Title', content='', target=''):
             content = content()
     if target:
         target = ' target="{}"'.format(target)
+    if escaped:
+        content = escape(content)
     return u'<a href="%s"%s>%s</a>' % (href, target, safe_unicode(content))